+++ setup.cfg
@@ -43,7 +43,7 @@ exclude =
 # Ignoring some errors in some files:
 per-file-ignores =
   # Enable `assert` keyword and magic numbers for tests:
-  tests/*.py: S101, E501, WPS226, WPS432
+  tests/*.py: S101, E501, WPS226, WPS432, WPS202
   sterces/cli.py: WPS211, WPS216
   sterces/entries/commands.py: WPS211, WPS216, WPS226
   sterces/foos.py: WPS234, WPS221
