+++ sterces/constants.py
@@ -1,12 +1,14 @@
 """Top level module constants in package sterces."""
 
-from types import MappingProxyType
+from pathlib import Path
 
 VERSION = "0.1.0"
-
-CONTEXT_SETTINGS = MappingProxyType({"help_option_names": ["-h", "--help"]})
 
 # actions
 ADD = "add"
 REMOVE = "remove"
 SHOW = "show"
+
+STERCES_DN = Path().home() / ".sterces"
+DEFAULT_DB_FN = str(STERCES_DN / "db.kdbx")
+DEFAULT_PWD_FN = str(STERCES_DN / ".ssapeek")
